Moved notification handlers out of the Compose namespace
